move files to folders ssr and csr-and-ssr
