+++ projects/schematics/src/modernize-app-migrated-from-68-to-221119/csr-and-ssr/update-angular-json-for-application-builder.ts
@@ -1,5 +1,5 @@
 import { Rule, SchematicContext, Tree } from '@angular-devkit/schematics';
-import { getWorkspace } from '../shared/utils/workspace-utils';
+import { getWorkspace } from '../../shared/utils/workspace-utils';
 
 /**
  * Updates the Angular configuration file to new Angular v17 standards.
