+++ projects/schematics/src/modernize-app-migrated-from-68-to-221119/csr-and-ssr/update-app-module.ts
