+++ projects/schematics/src/modernize-app-migrated-from-68-to-221119/ssr/update-ts-config-app.ts
